requirements.txt
mrcfile
numpy
pandas
pyserialem
scipy
tifffile
pyyaml
tqdm